+++ src/components/customers/CustomerList.vue
@@ -1,0 +1,16 @@
+<template>
+    <p>Lista</p>
+</template>
+<script>
+    export default {
+        name: "customerList",
+        data() {
+            return {
+
+            };
+        },
+        methods: {
+
+        }
+    };
+</script>