+++ src/main.js
@@ -1,0 +1,13 @@
+import '@babel/polyfill'
+import './styles/app.scss';
+import Vue from 'vue'
+import './plugins/vuetify'
+import App from './App.vue'
+import router from './router'
+
+Vue.config.productionTip = false;
+
+new Vue({
+  router,
+  render: h => h(App)
+}).$mount('#app');